fix tests

+++ src/anaconda/install.sh
@@ -104,6 +104,7 @@ if ! conda --version &> /dev/null ; then
 fi
 
 # Display a notice on conda when not running in GitHub Codespaces
+mkdir -p /usr/local/etc/vscode-dev-containers
 cat << 'EOF' > /usr/local/etc/vscode-dev-containers/conda-notice.txt
 When using "conda" from outside of GitHub Codespaces, note the Anaconda repository contains
 restrictions on commercial use that may impact certain organizations. See https://aka.ms/ghcs-conda
@@ -119,6 +120,8 @@ fi
 EOF
 )"
 
-echo "${notice_script}" | tee -a /etc/bash.bashrc >> /etc/zsh/zshrc
+if [ -f "/etc/zsh/zshrc" ]; then
+    echo "${notice_script}" | tee -a /etc/bash.bashrc >> /etc/zsh/zshrc
+fi
 
 echo "Done!"
